% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{configure-python}
\alias{configure-python}
\alias{validate_python}
\alias{configure_antspynet}
\alias{configure_python}
\alias{remove_conda}
\alias{ensure_rpymat}
\title{Install & Configure 'python' environment for 'RAVE'}
\usage{
validate_python(verbose = TRUE, env_name = NA)

configure_antspynet()

configure_python(python_ver = "3.11", verbose = TRUE)

remove_conda(ask = TRUE)

ensure_rpymat(env_name = NA)
}
\arguments{
\item{verbose}{whether to verbose messages}

\item{env_name}{'conda' environment to use; set to \code{NA} to verify
the default environment}

\item{python_ver}{python version; default is automatically determined. If
a specific python version is needed, please specify explicitly, for example,
\code{python_ver='3.8'}}

\item{ask}{whether to ask before resetting the 'python' environment}
}
\description{
Installs 'python' environment for 'RAVE'
}
\details{
Use \code{ravemanager::configure_python()} to install and configure
python environment using \code{miniconda}. The \code{conda} binary and
environment will be completely isolated. This means the installation will be
safe and it will not affect any existing configurations on the computer.

In this isolated environment, the following packages will be installed:
\code{numpy}, \code{scipy}, \code{pandas}, \code{h5py}, \code{jupyterlab},
\code{pynwb}, \code{mat73}, \code{mne}. You can always
add more \code{conda} packages via \code{rpymat::add_packages(...)} or
\code{pip} packages via \code{rpymat::add_packages(..., pip = TRUE)}.

To use the 'python' environment, please run \code{rpymat::ensure_rpymat()}
to activate in your current session. If you are running via
\code{'RStudio'}, open any 'python' script and use \code{'ctrl/cmd+enter'}
to run line-by-line. To switch from \code{R} to \code{python} mode,
using command \code{rpymat::repl_python()}

A \code{jupyterlab} will be automatically installed during the
configuration. To launch the \code{jupyterlab}, use
\code{rpymat::jupyter_launch()}

If you want to remove this \code{conda} environment, use R command
\code{rpymat::remove_conda()}. This procedure is absolutely safe and
will not affect your other installations.
}
\examples{

\dontrun{

# -------- Install & Configure python environment for RAVE --------
ravemanager::configure_python()

# Add conda packages
rpymat::add_packages("numpy")

# Add pip packages
rpymat::add_packages("nipy", pip = TRUE)

# -------- Activate RAVE-python environment --------
rpymat::ensure_rpymat()

# run script from a temporary file
f <- tempfile(fileext = ".py")
writeLines(c(
  "import numpy",
  "print(f'numpy installed: {numpy.__version__}')"
), f)
rpymat::run_script(f)

# run terminal command within the environment
rpymat::run_command("pip list")

# run python interactively, remember to use `exit` to exit
# python mode
rpymat::repl_python()



}

}
