% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{debug_info}
\alias{debug_info}
\title{Print 'RAVE' debugging information}
\usage{
debug_info(max_lines)
}
\arguments{
\item{max_lines}{maximum number of log entries to print}
}
\description{
This function will generate a report. Please include the report when you
file an issue. Useful for reporting issues to 'RAVE' developers.
}
