% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{find_packages_with_empty_files}
\alias{find_packages_with_empty_files}
\title{Find packages with empty files}
\usage{
find_packages_with_empty_files(lib = get_libpaths(check = TRUE))
}
\arguments{
\item{lib}{library path where packages are installed; default is set to
user library path.}
}
\value{
A list of packages (and files) containing empty files.
}
\description{
Check whether packages are installed correctly. In rare cases (possibly
network issues), packages downloaded contain empty files. This function
provides a method to check empty files in packages.
}
\examples{

find_packages_with_empty_files()

}
