% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_packages}
\alias{install_packages}
\alias{add_r_package}
\alias{add_py_package}
\title{Install/Update R or Python packages to RAVE environment}
\usage{
add_r_package(
  pkg,
  lib = get_libpaths(check = TRUE),
  repos = get_mirror(),
  type = getOption("pkgType"),
  ...,
  INSTALL_opts = "--no-lock"
)

add_py_package(pkg, method = c("pip", "conda"))
}
\arguments{
\item{pkg}{name of the package}

\item{repos, lib, type, INSTALL_opts, ...}{internally used}

\item{method}{whether to use \code{'pip'} or \code{'conda'}; default is
\code{'pip'}}
}
\value{
Nothing
}
\description{
Install/Update R or Python packages to RAVE environment
}
\examples{
\dontrun{


# ---- R --------------------------------------------------------
# Install R packages (CRAN, BioC, or RAVE's repository)
add_r_package("ravebuiltins")
add_r_package("rhdf5")

# Install from Github (github.com/dipterix/threeBrain)
add_r_package("dipterix/threeBrain")

# Install Github branch
add_r_package("dipterix/threeBrain@custom-electrode-geom")

# ---- Python ----------------------------------------------------

# Normal pypi packages
add_py_package("threebrainpy")

# Add through conda
add_py_package("fftw", method = "conda")


}


}
