% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{RAVE-install}
\alias{RAVE-install}
\alias{add_shortcuts}
\alias{finalize_installation}
\alias{clear_cache}
\alias{install}
\alias{install_dev}
\alias{update_rave}
\alias{upgrade_installer}
\title{Install or upgrade 'RAVE'}
\usage{
add_shortcuts()

finalize_installation(
  packages = NULL,
  upgrade = c("config-only", "ask", "always", "never", "data-only"),
  async = FALSE,
  ...
)

clear_cache()

install(
  allow_cache = FALSE,
  upgrade_manager = FALSE,
  finalize = TRUE,
  force = FALSE,
  python = FALSE,
  migrate_packages = FALSE,
  lib_path = NA,
  ...
)

install_dev(branch_name)

update_rave(
  allow_cache = FALSE,
  upgrade_manager = FALSE,
  finalize = TRUE,
  force = FALSE,
  python = FALSE,
  migrate_packages = FALSE,
  lib_path = NA,
  ...
)

upgrade_installer(reload = TRUE)
}
\arguments{
\item{packages}{packages to run finalizing installation scripts}

\item{upgrade}{upgrade type}

\item{async}{whether to execute finalizing installation scripts in other
processes}

\item{...}{passed to internal functions}

\item{allow_cache}{whether to allow cache; default is true}

\item{upgrade_manager}{whether to upgrade the installer (\code{ravemanager})
before updating other packages}

\item{finalize}{whether to run finalizing installation scripts}

\item{force}{whether to force updating packages even the installed have
the latest versions}

\item{python}{whether to install python; default is false}

\item{migrate_packages}{whether to migrate (copy) packages installed in
the system path to the user library; used as alternative option when
there are existing packages installed in the system library path that could
be hard to resolve or out-dated; default is \code{FALSE}}

\item{lib_path}{library path where 'RAVE' should be installed; default is
automatically determined.}

\item{branch_name}{for 'RAVE' developers, development branch to install}

\item{reload}{whether to reload \code{ravemanager} after installation;
default is true. This tries to load the upgraded \code{ravemanager}
without restarting the R session; however, this solution not always
works. In such case, restarting R session is always the solution.}
}
\value{
Nothing
}
\description{
Installs the newest version of 'RAVE' and its dependence
packages; executes the scripts to finalize installation to update
configuration files.
}
