% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{rave-profile}
\alias{rave-profile}
\alias{make_profile}
\alias{use_profile}
\title{Make a profile allowing portable \code{RAVE} configurations}
\usage{
make_profile(profile = shared_profile(), profile_name = "default")

use_profile(profile_name = "default", auto_install = TRUE)
}
\arguments{
\item{profile}{profile configurations}

\item{profile_name}{name of the profile}

\item{auto_install}{whether to automatically install RAVE with the
profile; default is \code{TRUE}}
}
\description{
Make a profile allowing portable \code{RAVE} configurations
}
\examples{
\dontrun{

# Make a template profile first. You may edit this profile
# to redirect paths to shared drive, SSD etc.
# Make sure you have the write permissions

make_profile()

# Test and install the profile
use_profile()

# Edit in your .Rprofile via `usethis::edit_r_profile()`
# add this line
ravemanager::use_profile(auto_install = FALSE)

}

}
