% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{ravemanager_version}
\alias{ravemanager_version}
\title{Get current 'RAVE' installer's version}
\usage{
ravemanager_version()
}
\value{
Returns the \code{ravemanager} version
}
\description{
Get current 'RAVE' installer's version
}
