% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial.R
\name{run_tutorials}
\alias{run_tutorials}
\title{Run tutorials}
\usage{
run_tutorials(topic = NULL, ...)
}
\arguments{
\item{topic}{integers of which topic to launch, leave it blank, then 'RAVE'
will ask you to select a topic}

\item{...}{other parameters to pass to \code{shiny::runApp()}}
}
\description{
Run tutorials
}
\examples{
\dontrun{

ravemanager::run_tutorials()

}
}
