% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall RAVE components}
\usage{
uninstall(components = c("cache", "python", "all"))
}
\arguments{
\item{components}{which component to remove, see example for choices.}
}
\description{
Remove cache, python, and/or all settings. Please be aware that
R, 'RStudio', and already installed R packages will not be uninstalled.
Please carefully read printed messages.
}
\examples{

if ( FALSE ) {

  # remove cache only
  ravemanager::uninstall("cache")

  # remove python environment
  ravemanager::uninstall("python")

  # remove all sample data, settings files
  ravemanager::uninstall("all")

}


}
