% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wasm-helper.R
\name{webr_view}
\alias{webr_view}
\title{Helper function to view \code{'htmlwidgets'} in 'WebAssembly'}
\usage{
webr_view(widget)
}
\arguments{
\item{widget}{\code{'htmlwidgets'} widget}
}
\value{
nothing
}
\description{
Helper function to view \code{'htmlwidgets'} in 'WebAssembly'
}
