% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{RAVE-install}
\alias{RAVE-install}
\alias{add_shortcuts}
\alias{finalize_installation}
\alias{install}
\alias{update_rave}
\alias{upgrade_installer}
\title{Install or upgrade 'RAVE'}
\usage{
add_shortcuts()

finalize_installation(
  packages = NULL,
  upgrade = c("config-only", "ask", "always", "never", "data-only"),
  async = FALSE,
  ...
)

install(
  nightly = FALSE,
  upgrade_manager = FALSE,
  finalize = TRUE,
  force = FALSE,
  ...
)

update_rave(
  nightly = FALSE,
  upgrade_manager = FALSE,
  finalize = TRUE,
  force = FALSE,
  ...
)

upgrade_installer()
}
\arguments{
\item{packages}{packages to run finalizing installation scripts}

\item{upgrade}{upgrade type}

\item{async}{whether to execute finalizing installation scripts in other
processes}

\item{...}{passed to internal functions, useful arguments include
\code{use_rspm} to whether use \code{'RSPM'} on \code{'Ubuntu'} (enabled by default)}

\item{nightly}{whether to install the nightly build}

\item{upgrade_manager}{whether to upgrade the installer (\code{ravemanager})
before updating other packages}

\item{finalize}{whether to run finalizing installation scripts}

\item{force}{whether to force updating packages even the installed have
the latest versions}
}
\value{
Nothing
}
\description{
Installs the newest version of 'RAVE' and its dependence
packages; executes the scripts to finalize installation to update
configuration files.
}
