% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{RAVE-install}
\alias{RAVE-install}
\alias{finalize_installation}
\alias{install}
\title{Install or upgrade 'RAVE'}
\usage{
finalize_installation(
  packages = NULL,
  upgrade = c("config-only", "ask", "always", "never", "data-only"),
  async = FALSE,
  ...
)

install(nightly = TRUE, upgrade_manager = TRUE)
}
\arguments{
\item{packages}{packages to run finalizing installation scripts}

\item{upgrade}{upgrade type}

\item{async}{whether to execute finalizing installation scripts in other
processes}

\item{...}{reserved for future use}

\item{nightly}{whether to install the nightly build}

\item{upgrade_manager}{whether to upgrade the installer (\code{ravemanager})
before updating other packages}
}
\value{
Nothing
}
\description{
Installs the newest version of 'RAVE' and its dependence
packages; executes the scripts to finalize installation to update
configuration files.
}
