% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system-requirements.R
\name{system_requirements}
\alias{system_requirements}
\title{Get system requirements for 'RAVE'}
\usage{
system_requirements(
  os = NULL,
  os_release = NULL,
  curl = Sys.which("curl"),
  sudo = FALSE,
  ...
)
}
\arguments{
\item{os, os_release}{operating system and release version, currently only
supports \code{'ubuntu'}, \code{'centos'}, \code{'redhat'}, \code{'opensuse'},
and \code{'sle'}; see
\url{https://github.com/rstudio/r-system-requirements#operating-systems}}

\item{curl}{the location of the curl binary on your system}

\item{sudo}{whether to pre-pend \code{'sudo'} to the commands}

\item{...}{reserved for future use}
}
\description{
Get system requirements for 'RAVE'
}
\examples{


if("remotes" \%in\% loadedNamespaces()) {
# Please check your operating system & version!!!

# =============== On Ubuntu Linux ===============

# Ubuntu 20
ravemanager::system_requirements("ubuntu", "20")

# Ubuntu 18
ravemanager::system_requirements("ubuntu", "18")

# Ubuntu 16
ravemanager::system_requirements("ubuntu", "16")

# =============== On Red Hat Enterprise Linux ===============

# Red Hat Enterprise Linux 8
ravemanager::system_requirements("redhat", "8")

# Red Hat Enterprise Linux 7
ravemanager::system_requirements("redhat", "7")

# =============== On CentOS ===============

# Red Hat Enterprise Linux 8
ravemanager::system_requirements("centos", "8")

# Red Hat Enterprise Linux 7
ravemanager::system_requirements("centos", "7")

# =============== On OpenSUSE ===============

# openSUSE 42.3
ravemanager::system_requirements("opensuse", "42")

# =============== On SUSE Linux Enterprise ===============

# SUSE Linux Enterprise 12.3
ravemanager::system_requirements("sle", "12")

}


}
