% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{version_info}
\alias{version_info}
\title{Print out 'RAVE' version information}
\usage{
version_info(vanilla = FALSE)
}
\arguments{
\item{vanilla}{whether to use vanilla packages in this function}
}
\description{
Print out 'RAVE' version information
}
